Fix compile build issues for Windows

+++ SuiteSparse/GPUQREngine/Include/GPUQREngine_BucketList.hpp
@@ -62,7 +62,8 @@ public:
     int VThead;              // Index of the first available entry in VTlist
 
     // Constructors
-    void *operator new(long unsigned int, BucketList* p)
+	//void *operator new(long unsigned int, BucketList* p)
+	void *operator new(size_t, BucketList* p)
     {
         return p;
     }
